backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pymupdf==1.23.8
openai>=1.6.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
email-validator==2.1.0
jinja2==3.1.2